% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charset_size.R
\name{charset_size}
\alias{charset_size}
\title{Character-Set Size}
\usage{
charset_size(x)
}
\arguments{
\item{x}{a character string of the password to test}
}
\value{
a numeric value of the charset size
}
\description{
Determine the Charset size based on what combination of characters has been used in the password
}
