% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_strength.R
\name{password_strength}
\alias{password_strength}
\title{Password Strength}
\usage{
password_strength(x)
}
\arguments{
\item{x}{a character string of the password to test}
}
\value{
a list of two elements
\itemize{
\item ENTROPY; a numeric value for the password entropy
\item STRENGTH; a strength rating of the password (Very Weak, Weak, Moderate, Strong, or Very Strong)
}
}
\description{
Test the strength of a password
}
